webService/webService/requirements.txt
django
mysqlclient
django-rest-framework
django-cors-headers
django-crispy-forms
requests
